backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
httpx==0.27.0
tiktoken==0.7.0
langdetect==1.0.9
deep-translator==1.11.4
argostranslate==1.9.6
python-dotenv==1.0.1
slowapi==0.1.9
aiosqlite==0.20.0
greenlet==3.0.3
fastapi-mail==1.4.1
authlib==1.3.0
itsdangerous==2.1.2
jinja2==3.1.4
